Add ParallelExecutionError to fatal hierarchy
